✨[FEAT]: #18 - 뷰 그림자 익스텐션 구현 및 홈 네비게이션 뷰 그림자 적용

+++ ChaRo-iOS/ChaRo-iOS/Source/Views/Cells/TVC/HomeAnimationTVC.swift
@@ -42,7 +42,7 @@ class HomeAnimationTVC: UITableViewCell {
 
         if scrollView.contentOffset.x > 0{
                if scrollView.contentOffset.x < 1000 {
-                carMoveConstraint.constant = (scrollView.contentOffset.x - sideMargin)/pageCount
+                carMoveConstraint.constant = (scrollView.contentOffset.x - sideMargin)/CGFloat(pageCount)
                } else {
                 carMoveConstraint.constant = originalCarConstant
                }
